fixed objectcoordinates taking doubles instead of integers

+++ stargazer/src/main/java/com/omerygouw/stargazer/DTO/ObjectCoordinates.java
@@ -3,5 +3,5 @@ package com.omerygouw.stargazer.DTO;
 import lombok.Builder;
 
 @Builder
-public record ObjectCoordinates(Double azimuth, Double altitude) {
+public record ObjectCoordinates(Integer azimuth, Integer altitude) {
 }
